requirements.txt

python-dotenv
python-binance
pandas
numpy
joblib
xgboost
ta
scikit-learn
matplotlib
flask
